+++ src/components/SearchResources/SearchResources.js
@@ -13,8 +13,8 @@ import {
   bibleSubjects,
   obsSubjects,
 } from '../../config/materials';
-import { defaultCard, server } from '../../config/base';
-import { getXY } from '../../core/matrix';
+import { defaultCard, server, columns } from '../../config/base';
+import { getXY } from 'resource-workspace-rcl';
 import { getUniqueResources } from '../../helper';
 
 import { MenuItem, Menu } from '@material-ui/core';
@@ -40,18 +40,29 @@ function SearchResources({ anchorEl, onClose, open }) {
   const uniqueResources = getUniqueResources(appConfig, resourcesApp);
 
   const handleAddMaterial = (item) => {
-    const pos = getXY(appConfig);
-    setAppConfig((prev) =>
-      prev.concat({ ...defaultCard, x: pos.x, y: pos.y, i: item.name })
-    );
-    onClose();
+    setAppConfig((prev) => {
+      const next = { ...prev };
+      for (let k in next) {
+        const pos = getXY(appConfig[k], columns[k], defaultCard[k].h, defaultCard[k].w);
+        next[k] = next[k].concat({
+          ...defaultCard[k],
+          x: pos.x,
+          y: pos.y,
+          i: item.name,
+        });
+      }
+      return next;
+    });
+    setTimeout(function () {
+      window.scrollTo(0, 10000);
+    }, 1000);
   };
 
   useEffect(() => {
     axios
       .get(
         server +
-          '/api/catalog/v5/search?sort=lang,title&owner=' +
+          '/api/catalog/v5/search?limit=1000&sort=lang,title&owner=' +
           owners.join(',') +
           '&lang=' +
           langs.join(',') +
@@ -106,7 +117,7 @@ function SearchResources({ anchorEl, onClose, open }) {
         );
       }
     });
-
+  const emptyMenuItems = <p className={classes.divider}>{t('No_resources')}</p>;
   return (
     <Menu
       color="transparent"
@@ -115,7 +126,7 @@ function SearchResources({ anchorEl, onClose, open }) {
       open={open}
       onClose={onClose}
     >
-      {menuItems}
+      {menuItems.length !== 0 ? menuItems : emptyMenuItems}
     </Menu>
   );
 }