+++ src/components/WorkspaceManager/WorkspaceManager.js
@@ -4,12 +4,7 @@ import { useTranslation } from 'react-i18next';
 
 import { AppContext, ReferenceContext } from '../../context';
 
-import {
-  defaultTplBible,
-  defaultTplOBS,
-  defaultBibleReference,
-  defaultOBSReference,
-} from '../../config/base';
+import { resetWorkspace } from '../../helper';
 import { MenuItem, MenuList } from '@material-ui/core';
 
 function WorkspaceManager({ onClose }) {
@@ -26,48 +21,22 @@ function WorkspaceManager({ onClose }) {
   } = useContext(ReferenceContext);
 
   const { t } = useTranslation();
-  const workspaceType = bookId === 'obs' ? 'obs' : 'bible';
+
   const handleReset = () => {
-    const oldAppConfig = JSON.parse(localStorage.getItem('appConfig'));
-    switch (workspaceType) {
-      case 'bible':
-        const bibleAppConfig = {
-          ...oldAppConfig,
-          [workspaceType]: defaultTplBible[currentLanguage],
-        };
-        localStorage.setItem('appConfig', JSON.stringify(bibleAppConfig));
-        setAppConfig(defaultTplBible[currentLanguage]);
-        goToBookChapterVerse(
-          defaultBibleReference[currentLanguage].bookId,
-          defaultBibleReference[currentLanguage].chapter,
-          defaultBibleReference[currentLanguage].verse
-        );
-        break;
-
-      case 'obs':
-        const obsAppConfig = {
-          ...oldAppConfig,
-          [workspaceType]: defaultTplOBS[currentLanguage],
-        };
-        localStorage.setItem('appConfig', JSON.stringify(obsAppConfig));
-        setAppConfig(defaultTplOBS[currentLanguage]);
-        goToBookChapterVerse(
-          defaultOBSReference[currentLanguage].bookId,
-          defaultOBSReference[currentLanguage].chapter,
-          defaultOBSReference[currentLanguage].verse
-        );
-        break;
-
-      default:
-        break;
-    }
+    resetWorkspace({
+      bookId,
+      setAppConfig,
+      goToBookChapterVerse,
+      currentLanguage,
+      resetAll: false,
+    });
     onClose();
   };
 
   return (
     <MenuList>
       <MenuItem divider={true} onClick={handleReset}>
-        {t('Reset')}
+        {t('Reset_cards')}
       </MenuItem>
     </MenuList>
   );