+++ src/context/AppContext.js
@@ -4,25 +4,21 @@ import { ResourcesContextProvider } from 'scripture-resources-rcl';
 
 import { useTranslation } from 'react-i18next';
 import { ReferenceContext } from './ReferenceContext';
-import { getResources, getBookList, checkLSVal } from '../helper';
+import { getResources, getBookList, checkLSVal, getLayoutType } from '../helper';
 import {
-  server,
   defaultTplBible,
   defaultTplOBS,
   languages,
   bibleList,
+  server,
 } from '../config/base';
 
 export const AppContext = React.createContext();
 
 const _currentLanguage = checkLSVal('i18nextLng', languages[0]);
 const _resourcesApp = checkLSVal('resourcesApp', [], 'object');
-
-const _loadIntro = checkLSVal('loadIntro', true, 'boolean');
-
+const _startDialog = checkLSVal('startDialog', true, 'boolean');
 const _fontSize = parseInt(localStorage.getItem('fontSize'));
-
-const config = { server };
 
 export function AppContextProvider({ children }) {
   const {
@@ -31,7 +27,6 @@ export function AppContextProvider({ children }) {
   } = useContext(ReferenceContext);
 
   const [currentLanguage, setCurrentLanguage] = useState(_currentLanguage);
-
   const [appConfig, setAppConfig] = useState(
     () =>
       checkLSVal(
@@ -45,6 +40,12 @@ export function AppContextProvider({ children }) {
       )[referenceSelected.bookId === 'obs' ? 'obs' : 'bible']
   );
 
+  const [breakpoint, setBreakpoint] = useState({ name: 'lg', cols: 12 });
+  /** TODO Create ResourceContext
+   * 1. Get information about resources ( like available bookId) from /Chapter  - content.resources.project.
+   * 2. Put all states about resources in ResourceContext.
+   * 3. Maybe make availableBookList in ResourceContext
+   */
   const [resourcesApp, setResourcesApp] = useState(_resourcesApp);
   const _resourceLinks = getResources(appConfig, resourcesApp);
   const [resourceLinks, setResourceLinks] = useState(_resourceLinks);
@@ -54,10 +55,10 @@ export function AppContextProvider({ children }) {
   const [showErrorReport, setShowErrorReport] = useState(false);
   const [errorFile, setErrorFile] = useState('');
   const [fontSize, setFontSize] = useState(_fontSize ? _fontSize : 100);
-  const [loadIntro, setLoadIntro] = useState(_loadIntro);
-
+  const [loadIntro, setLoadIntro] = useState(false);
+  const [openStartDialog, setOpenStartDialog] = useState(_startDialog);
   const [openMainMenu, setOpenMainMenu] = useState(false);
-
+  const config = { server };
   const { t } = useTranslation();
 
   useEffect(() => {
@@ -65,8 +66,17 @@ export function AppContextProvider({ children }) {
   }, [fontSize]);
 
   useEffect(() => {
+    const type = getLayoutType(appConfig.lg);
+    const newType = referenceSelected.bookId === 'obs' ? 'obs' : 'bible';
+    if (type !== newType) {
+      setAppConfig(JSON.parse(localStorage.getItem('appConfig'))[newType]);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [referenceSelected.bookId]);
+
+  useEffect(() => {
     setResourceLinks(getResources(appConfig, resourcesApp));
-  }, [appConfig, resourcesApp]);
+  }, [appConfig, resourcesApp, breakpoint]);
 
   useEffect(() => {
     setNewBookList(getBookList(bibleList, t), true);
@@ -80,15 +90,20 @@ export function AppContextProvider({ children }) {
   useEffect(() => {
     localStorage.setItem('loadIntro', loadIntro);
   }, [loadIntro]);
+  useEffect(() => {
+    localStorage.setItem('startDialog', openStartDialog);
+  }, [openStartDialog]);
 
   const value = {
     state: {
       appConfig,
+      breakpoint,
       currentLanguage,
       errorFile,
       fontSize,
       loadIntro,
       openMainMenu,
+      openStartDialog,
       resourceLinks,
       resourcesApp,
       resources,
@@ -99,11 +114,13 @@ export function AppContextProvider({ children }) {
     },
     actions: {
       setAppConfig,
+      setBreakpoint,
       setCurrentLanguage,
       setErrorFile,
       setFontSize,
       setLoadIntro,
       setOpenMainMenu,
+      setOpenStartDialog,
       setResourceLinks,
       setResourcesApp,
       setResources,