+++ src/docs/CHANGELOG.md
@@ -4,6 +4,38 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-10-20
+
+### Added (for new features)
+- Language selection on first boot
+- Scrolldown when new card added
+- Catch error with reset config
+- Shortcut for clear localstorage (ctrl+shift+L)
+- Migration mechanizm [#203](https://github.com/texttree/bsa/issues/203)
+- Copy reference to clipboard [#218](https://github.com/texttree/bsa/issues/218)
+- Tajik localisation
+- More language support
+
+### Changed (for changes in existing functionality)
+- User guide extended for OBS
+- User guide update about synchronization cards
+- When switched the Bible or OBS mode, after pressing the back button, the reference switched, but the resources did not
+- Do not close the add resources menu automatically
+- Move abbr to translation [#206](https://github.com/texttree/bsa/issues/206)
+- Update resource-workspace-rcl [#207](https://github.com/texttree/bsa/issues/207)
+- For ru project set default ref tit 1:1 [#208](https://github.com/texttree/bsa/issues/208)
+- Color for current verse
+- Don't render all current cards when another one is loaded or deleted [#211](https://github.com/texttree/bsa/issues/211)
+- Color of theme in manifest
+
+### Fixed (for any bug fixes)
+- Highlight the first verse at the first launch
+- Remove localization in pl and ukr, cause it breaks the app
+- Remove the UEB, replace it with ULT
+- Fix empty place when all materials added
+- Order of NT books [#205](https://github.com/texttree/bsa/issues/205)
+- Show OBS header when 1st verse opened
+- Scroll to verse when reset layout
 
 ## [1.2.0] - 2021-10-05
 
@@ -17,7 +49,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security (in case of vulnerabilities)
 - Updating dependencies
-
 
 ## [1.1.1] - 2021-09-29